move constants to constants folder
previous src/constants.js caused eslint errors with absolute imports. Apparently `constants` is a deprecated node module.
