+++ src/containers/MathObjects/Folder/components/CollapsedIndicator.js
@@ -3,7 +3,7 @@ import { Icon } from 'antd'
 import styled from 'styled-components'
 import PropTypes from 'prop-types'
 import SubtleButton from 'components/SubtleButton'
-import { theme } from 'theme'
+import { theme } from 'constants/theme'
 
 const CollapseIndicatorButton = styled(SubtleButton)`
   padding-left: 2px;
