+++ src/index.js
@@ -6,7 +6,7 @@ import './index.css'
 import App from './App'
 import registerServiceWorker from './registerServiceWorker'
 
-import theme from './theme'
+import theme from './constants/theme'
 import { ThemeProvider } from 'styled-components'
 
 const target = document.querySelector('#root')
